+++ frontend/src/components/FeatureCardsSection/FeatureCardsSection.tsx
@@ -1,13 +1,10 @@
-import React from 'react';
-import { Box, Typography } from '@mui/material';
-import { styled } from '@mui/material';
-import { FeatureCards } from './components';
+import { Box, Typography, styled } from '@mui/material';
+import FeatureCardGrid from './components/FeatureCardGrid';
 
 const ToolsContainer = styled(Box)(({ theme }) => ({
   background: theme.palette.background.default,
   margin: '0 auto',
   padding: theme.spacing(8, 2),
-
   [theme.breakpoints.up('sm')]: {
     padding: theme.spacing(8, 4),
   },
@@ -15,25 +12,34 @@ const ToolsContainer = styled(Box)(({ theme }) => ({
 
 const HeaderContainer = styled('div')(({ theme }) => ({
   textAlign: 'center',
-  marginBottom: theme.spacing(6), 
+  marginTop: theme.spacing(8),
+  marginBottom: theme.spacing(16),
 }));
 
-const Subtitle = styled(Typography)(({ theme }) => ({
-  marginTop: theme.spacing(2), 
+const HeaderTitle = styled(Typography)(({ theme }) => ({
+  ...theme.typography.h3,                 
+  color: theme.palette.text.primary,
+}));
+
+const HeaderSubtitle = styled(Typography)(({ theme }) => ({
+  ...theme.typography.subtitle1,           
+  fontWeight: theme.typography.fontWeightRegular, 
+  color: theme.palette.text.secondary,
+  marginTop: theme.spacing(6),
 }));
 
 const FeatureCardsSection = () => {
   return (
     <ToolsContainer>
       <HeaderContainer>
-        <Typography variant="h4">
+        <HeaderTitle>
           All-in-One Tools to Simplify Your Journey
-        </Typography>
-        <Subtitle variant="body1">
+        </HeaderTitle>
+        <HeaderSubtitle>
           Let's plan smarter: from viewing a home to securing a loan and optimizing your super
-        </Subtitle>
+        </HeaderSubtitle>
       </HeaderContainer>
-      <FeatureCards />
+      <FeatureCardGrid />
     </ToolsContainer>
   );
 };
