--- frontend/src/components/FeatureCardsSection/components/components/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FeatureCardGrid';
