--- frontend/src/components/FeatureCardsSection/components/index.ts
@@ -1,1 +1,0 @@
-export { default as FeatureCards } from './FeatureCards';
